refactor: 互換 re-export シムを削除して本体パスに統一
src/adapters/initAdapter.ts / src/utils/apiBridge.ts / src/utils/loginPrompt.ts
は移動先への re-export だけのシムだったため削除し、参照元 28 ファイルを
本体パス (@/composables/useInitAdapter, @/core/apiBridge,
@/composables/useLoginPrompt) に直接リンクするよう更新。
