--- src/adapters/initAdapter.ts
@@ -1,7 +1,0 @@
-// Moved to @/composables/useInitAdapter.ts — re-export for compatibility
-export {
-  destroyAdapter,
-  destroyAllAdapters,
-  type InitAdapterResult,
-  initAdapterFor,
-} from '@/composables/useInitAdapter'
