+++ src/components/common/NoteMoreMenu.vue
@@ -8,6 +8,7 @@ import {
 } from '@/aiscript/plugin-api'
 import { useCommandStore } from '@/commands/registry'
 import { useAccountMode } from '@/composables/useAccountMode'
+import { showLoginPrompt } from '@/composables/useLoginPrompt'
 import { useMultiAccountAdapters } from '@/composables/useMultiAccountAdapters'
 import { useConfirm } from '@/stores/confirm'
 import { useDeckStore } from '@/stores/deck'
@@ -15,7 +16,6 @@ import { usePrompt } from '@/stores/prompt'
 import { useToast } from '@/stores/toast'
 import { useWindowsStore } from '@/stores/windows'
 import { AppError } from '@/utils/errors'
-import { showLoginPrompt } from '@/utils/loginPrompt'
 import { commands, unwrap } from '@/utils/tauriInvoke'
 import { isSafeUrl } from '@/utils/url'
 import PopupMenu from './PopupMenu.vue'
