+++ src/components/window/NoteDetailContent.vue
@@ -7,7 +7,6 @@ import {
   useTemplateRef,
   watch,
 } from 'vue'
-import { initAdapterFor } from '@/adapters/initAdapter'
 import type {
   NormalizedNote,
   NoteReaction,
@@ -22,6 +21,7 @@ import type {
   NoteTreeNode,
 } from '@/components/common/MkNoteTree.vue'
 import MkNoteTree from '@/components/common/MkNoteTree.vue'
+import { initAdapterFor } from '@/composables/useInitAdapter'
 import { commands, unwrap } from '@/utils/tauriInvoke'
 
 const MkPostForm = defineAsyncComponent(
