+++ src/composables/usePostFormState.ts
@@ -1,5 +1,4 @@
 import { computed, nextTick, ref, shallowRef, watch } from 'vue'
-import { initAdapterFor } from '@/adapters/initAdapter'
 import type {
   NormalizedNote,
   NoteVisibility,
@@ -20,6 +19,7 @@ import {
   saveDraft,
 } from '@/composables/useDrafts'
 import { useFileAttachment } from '@/composables/useFileAttachment'
+import { initAdapterFor } from '@/composables/useInitAdapter'
 import {
   deleteMemo,
   ensureMemosLoaded,
