--- src/utils/apiBridge.ts
@@ -1,2 +1,0 @@
-// Moved to @/core/apiBridge.ts — re-export for compatibility
-export { destroyApiBridge, initApiBridge } from '@/core/apiBridge'
