--- src/utils/loginPrompt.ts
@@ -1,2 +1,0 @@
-// Moved to @/composables/useLoginPrompt.ts — re-export for compatibility
-export { showLoginPrompt } from '@/composables/useLoginPrompt'
